feat(navbar): separate into smaller component for navbar
