+++ src/components/navbar/NavbarBrand.vue
@@ -1,0 +1,5 @@
+<template>
+  <router-link to="/" class="navbar-brand">
+    <img src="/src/assets/img/weblogo.png" alt="Logo" width="30" height="24" loading="lazy" />
+  </router-link>
+</template>
